LPS-92527 Move classes of com.liferay.portal.dao.orm package to module
