+++ modules/apps/portal/portal-dao-test/src/testIntegration/java/com/liferay/portal/dao/orm/test/SQLEscapedConcatTest.java
@@ -12,14 +12,17 @@
  * details.
  */
 
-package com.liferay.portal.dao.orm;
+package com.liferay.portal.dao.orm.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.dao.orm.common.SQLTransformer;
 import com.liferay.portal.kernel.dao.db.DB;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.test.rule.AssumeTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.util.PropsValues;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -27,75 +30,64 @@ import java.sql.ResultSet;
 
 import org.junit.AfterClass;
 import org.junit.Assert;
+import org.junit.Assume;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
- * @author Cristina González
+ * @author Michael Bowerman
  */
-public class SQLSubstrTest {
+@RunWith(Arquillian.class)
+public class SQLEscapedConcatTest {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
-		new LiferayIntegrationTestRule();
+		new AggregateTestRule(
+			new AssumeTestRule("assume"), new LiferayIntegrationTestRule());
+
+	public static void assume() {
+		String jdbcDefaultURL = PropsValues.JDBC_DEFAULT_URL;
+
+		Assume.assumeTrue(
+			jdbcDefaultURL.contains("mysql") ||
+			jdbcDefaultURL.contains("mariadb"));
+	}
 
 	@BeforeClass
 	public static void setUpClass() throws Exception {
 		_db = DBManagerUtil.getDB();
 
 		_db.runSQL(
-			"create table SQLSubstrTest (id LONG not null primary key, data " +
-				"VARCHAR(10) null)");
+			"create table SQLConcatTest (data VARCHAR(10) not null primary " +
+				"key)");
 
-		_db.runSQL("insert into SQLSubstrTest values (1, 'EXAMPLE')");
+		_db.runSQL("insert into SQLConcatTest values ('test')");
 	}
 
 	@AfterClass
 	public static void tearDownClass() throws Exception {
-		_db.runSQL("drop table SQLSubstrTest");
+		_db.runSQL("drop table SQLConcatTest");
 	}
 
 	@Test
-	public void testSubstr() throws Exception {
-		String sql = _db.buildSQL(
-			"select SUBSTR(data, 3, 2) from SQLSubstrTest where id = 1");
+	public void testConcatWithEscapedQuotes() throws Exception {
+		try (Connection con = DataAccess.getConnection();
+			PreparedStatement ps = con.prepareStatement(
+				SQLTransformer.transform(
+					"select CONCAT('This is a \\'', data, '\\' for escaped " +
+						"quotes') from SQLConcatTest"));
+			ResultSet rs = ps.executeQuery()) {
 
-		sql = SQLTransformer.transform(sql);
+			Assert.assertTrue(rs.next());
 
-		try (Connection connection = DataAccess.getConnection();
-			PreparedStatement preparedStatement = connection.prepareStatement(
-				sql)) {
+			Assert.assertEquals(
+				"This is a 'test' for escaped quotes", rs.getString(1));
 
-			ResultSet resultSet = preparedStatement.executeQuery();
-
-			Assert.assertNotNull(resultSet.next());
-
-			String substring = resultSet.getString(1);
-
-			Assert.assertEquals("AM", substring);
-		}
-	}
-
-	@Test
-	public void testSubstrStart() throws Exception {
-		String sql = _db.buildSQL(
-			"select SUBSTR(data, 1, 3) from SQLSubstrTest where id = 1");
-
-		sql = SQLTransformer.transform(sql);
-
-		try (Connection connection = DataAccess.getConnection();
-			PreparedStatement ps = connection.prepareStatement(sql)) {
-
-			ResultSet rs = ps.executeQuery();
-
-			Assert.assertNotNull(rs.next());
-
-			String substring = rs.getString(1);
-
-			Assert.assertEquals("EXA", substring);
+			Assert.assertFalse(rs.next());
 		}
 	}
