+++ modules/apps/portal/portal-dao-test/src/testIntegration/java/com/liferay/portal/dao/orm/test/SQLNullDateTest.java
@@ -12,79 +12,73 @@
  * details.
  */
 
-package com.liferay.portal.dao.orm;
+package com.liferay.portal.dao.orm.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.dao.orm.common.SQLTransformer;
 import com.liferay.portal.kernel.dao.db.DB;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.AssumeTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.util.PropsValues;
 
 import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.Statement;
 
 import org.junit.AfterClass;
-import org.junit.Assert;
-import org.junit.Assume;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
- * @author Michael Bowerman
+ * @author Adolfo Pérez
  */
-public class SQLEscapedConcatTest {
+@RunWith(Arquillian.class)
+public class SQLNullDateTest {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new AssumeTestRule("assume"), new LiferayIntegrationTestRule());
-
-	public static void assume() {
-		String jdbcDefaultURL = PropsValues.JDBC_DEFAULT_URL;
-
-		Assume.assumeTrue(
-			jdbcDefaultURL.contains("mysql") ||
-			jdbcDefaultURL.contains("mariadb"));
-	}
+		new LiferayIntegrationTestRule();
 
 	@BeforeClass
 	public static void setUpClass() throws Exception {
 		_db = DBManagerUtil.getDB();
 
 		_db.runSQL(
-			"create table SQLConcatTest (data VARCHAR(10) not null primary " +
-				"key)");
+			"create table SQLNullDateTest1 (id LONG not null primary key, " +
+				"date_ DATE null)");
 
-		_db.runSQL("insert into SQLConcatTest values ('test')");
+		_db.runSQL("insert into SQLNullDateTest1 (id) values (1)");
+
+		_db.runSQL("create table SQLNullDateTest2 (id LONG not null)");
+
+		_db.runSQL("insert into SQLNullDateTest2 (id) values (1)");
 	}
 
 	@AfterClass
 	public static void tearDownClass() throws Exception {
-		_db.runSQL("drop table SQLConcatTest");
+		_db.runSQL("drop table SQLNullDateTest1");
+
+		_db.runSQL("drop table SQLNullDateTest2");
 	}
 
 	@Test
-	public void testConcatWithEscapedQuotes() throws Exception {
-		try (Connection con = DataAccess.getConnection();
-			PreparedStatement ps = con.prepareStatement(
-				SQLTransformer.transform(
-					"select CONCAT('This is a \\'', data, '\\' for escaped " +
-						"quotes') from SQLConcatTest"));
-			ResultSet rs = ps.executeQuery()) {
+	public void testNullDate() throws Exception {
+		try (Connection connection = DataAccess.getConnection();
+			Statement statement = connection.createStatement()) {
 
-			Assert.assertTrue(rs.next());
+			String sql =
+				"(select date_ from SQLNullDateTest1) union all (select " +
+					"[$NULL_DATE$] from SQLNullDateTest2)";
 
-			Assert.assertEquals(
-				"This is a 'test' for escaped quotes", rs.getString(1));
+			ResultSet resultSet = statement.executeQuery(
+				SQLTransformer.transform(sql));
 
-			Assert.assertFalse(rs.next());
+			resultSet.close();
 		}
 	}
